+++ matrix-sdk/src/main/java/org/matrix/androidsdk/call/MXCallsManager.java
@@ -146,7 +146,9 @@ public class MXCallsManager {
             }
         });
 
-        refreshTurnServer();
+        if (isSupported()) {
+            refreshTurnServer();
+        }
     }
 
     /**